refactor: remove v1 folder
