+++ packages/backend/src/api.controller.spec.ts
@@ -1,6 +1,6 @@
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppModule } from '../app.module';
+import { AppModule } from './app.module';
 import { APIController } from './api.controller';
 
 describe('V1 ApiController', () => {
