+++ packages/backend/src/app.module.ts
@@ -1,21 +1,16 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { RouterModule } from '@nestjs/core';
-import { ApiV1Module } from './v1/api.module';
+import { APIController } from './api.controller';
+import { UsersModule } from './users/users.module';
 
 @Module({
+  controllers: [APIController],
   imports: [
     ConfigModule.forRoot({
       envFilePath: process.env.NODE_ENV === 'development' ? '.env.development.template' : '.env.production',
       isGlobal: true,
     }),
-    ApiV1Module,
-    RouterModule.register([
-      {
-        path: 'api/v1',
-        module: ApiV1Module,
-      },
-    ]),
+    UsersModule,
   ],
 })
 export class AppModule {}
