--- packages/backend/src/v1/api.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { APIController } from './api.controller';
-
-@Module({
-  controllers: [APIController],
-})
-export class ApiV1Module {}
